refactor(query): replace union-find with relabeled groups and tombstones
Dynamic SCC maintenance and atomic batch publication are unchanged; only
the staleness mechanism is simplified:

- A merge now eagerly relabels member files to a fresh group id and
  removes the merged groups, so a stored group id is always live and no
  find/path-compression layer is needed.
- Sealed constituents are retracted with an atomic tombstone flag in the
  same critical section that relabels their members, replacing the
  root-canonicity check; `is_current` becomes a lock-free flag read.
- The cycle region for a closing edge is computed with per-candidate
  forward searches instead of building a reverse edge index.
- The pdg tests move to pdg_tests.rs unchanged in coverage.
