+++ crates/tinymist-query/src/adt/mod.rs
@@ -1,3 +1,2 @@
 pub mod revision;
-pub(crate) mod union_find;
 pub use tinymist_analysis::adt::*;
